feat: new api
BREAKING CHANGE: added new api
