+++ lib/design_system/components/primary_button.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class PrimaryButton extends StatelessWidget {
+  final String text;
+  final Function() onPressed;
+
+  final bool isActive;
+
+  const PrimaryButton({
+    Key? key,
+    required this.text,
+    required this.onPressed,
+    this.isActive = true,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return TextButton(
+      style: TextButton.styleFrom(
+        fixedSize: const Size(350, 70),
+        primary: Colors.white,
+        backgroundColor:
+            isActive ? Theme.of(context).colorScheme.secondary : Colors.grey,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(50),
+        ),
+      ),
+      onPressed: isActive ? onPressed : () {},
+      child: Text(
+        text,
+        style: const TextStyle(
+          fontSize: 22,
+        ),
+        textAlign: TextAlign.center,
+      ),
+    );
+  }
+}
