+++ lib/external/errors/food_empty_error.dart
@@ -1,0 +1,8 @@
+import '../../infra/datasources/errors/ifoodsource_error.dart';
+
+class FoodEmptyError implements IFoodSourceError {
+  @override
+  ErrorType getError() {
+    return ErrorType.http;
+  }
+}
