+++ lib/external/errors/food_null_error.dart
@@ -1,0 +1,8 @@
+import '../../infra/datasources/errors/ifoodsource_error.dart';
+
+class FoodNullError implements IFoodSourceError {
+  @override
+  ErrorType getError() {
+    return ErrorType.http;
+  }
+}
