+++ lib/infra/datasources/errors/ifoodsource_error.dart
@@ -1,0 +1,5 @@
+enum ErrorType { http }
+
+abstract class IFoodSourceError {
+  ErrorType getError();
+}
