+++ lib/infra/datasources/ifoodsource.dart
@@ -1,0 +1,12 @@
+import 'package:dartz/dartz.dart';
+
+import '../../domain/features/entities/food.dart';
+import 'errors/ifoodsource_error.dart';
+
+abstract class IFoodSource {
+  Future<Either<List<Food>, IFoodSourceError>> fetchFoods();
+
+  Future<Either<List<Food>, IFoodSourceError>> searchFood(String name);
+
+  Future<Either<Food, IFoodSourceError>> fetchFoodById(int id);
+}
