+++ pubspec.yaml
@@ -29,6 +29,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  dartz: ^0.10.1
   http: ^0.13.4
   cupertino_icons: ^1.0.2
   get: ^4.6.1
